@@ -488,7 +488,7 @@ func TestAuditLog(t *testing.T) {
 }
 
 var responseBodyWriters = map[string]func(tx *Transaction, body string) (*types.Interruption, int, error){
-	"WriteResponsequestBody": func(tx *Transaction, body string) (*types.Interruption, int, error) {
+	"WriteResponseBody": func(tx *Transaction, body string) (*types.Interruption, int, error) {
 		return tx.WriteResponseBody([]byte(body))
 	},
 	"ReadResponseBodyFromKnownLen": func(tx *Transaction, body string) (*types.Interruption, int, error) {
@@ -977,8 +977,9 @@ func TestMultipleCookiesWithSpaceBetweenThem(t *testing.T) {
 
 func collectionValues(t *testing.T, col collection.Collection) []string {
 	t.Helper()
-	var values []string
-	for _, v := range col.FindAll() {
+	all := col.FindAll()
+	values := make([]string, 0, len(all))
+	for _, v := range all {
 		values = append(values, v.Value())
 	}
 	return values
@@ -1126,7 +1127,7 @@ func TestTransactionSyncPool(t *testing.T) {
 			ID_: 1234,
 		},
 	})
-	for i := 0; i < 1000; i++ {
+	for i := range 1000 {
 		if err := tx.Close(); err != nil {
 			t.Fatal(err)
 		}
