@@ -16,6 +16,26 @@ import (
 	"github.com/corazawaf/coraza/v3/internal/memoize"
 )
 
+// Description:
+// Performs case-insensitive pattern matching like @pm but loads keywords from file(s).
+// Each line in the file represents one keyword. Lines starting with # are treated as comments
+// and empty lines are ignored. Uses the Aho-Corasick algorithm for efficient matching.
+// Also available as @pmf (shorthand alias).
+//
+// Arguments:
+// File path(s) containing keywords, one per line. Multiple files can be specified space-separated.
+//
+// Returns:
+// true if any keyword from the file(s) is found in the input, false otherwise
+//
+// Example:
+// ```
+// # Block user agents from denylist file
+// SecRule REQUEST_HEADERS:User-Agent "@pmFromFile /path/to/denylist.txt" "id:172,deny,log"
+//
+// # Multiple files with shorthand alias
+// SecRule ARGS "@pmf badwords.txt sqli-patterns.txt" "id:173,deny"
+// ```
 func newPMFromFile(options plugintypes.OperatorOptions) (plugintypes.Operator, error) {
 	filepath := options.Arguments
 
@@ -45,7 +65,7 @@ func newPMFromFile(options plugintypes.OperatorOptions) (plugintypes.Operator, e
 		DFA:                  false,
 	})
 
-	m, _ := memoize.Do(strings.Join(options.Path, ",")+filepath, func() (interface{}, error) { return builder.Build(lines), nil })
+	m, _ := memoize.Do(strings.Join(options.Path, ",")+filepath, func() (any, error) { return builder.Build(lines), nil })
 
 	return &pm{matcher: m.(ahocorasick.AhoCorasick)}, nil
 }
