@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 //go:build !tinygo
-// +build !tinygo
 
 package strings
 
@@ -83,9 +82,33 @@ func TestMaybeRemoveQuotes(t *testing.T) {
 	}
 }
 
+func TestUnescapeQuotedString(t *testing.T) {
+	tests := []struct {
+		input string
+		want  string
+	}{
+		{input: ``, want: ``},
+		{input: `hello`, want: `hello`},
+		{input: `\"`, want: `"`},
+		{input: `\\`, want: `\\`},
+		{input: `@contains \"`, want: `@contains "`},
+		{input: `@rx C:\\`, want: `@rx C:\\`},
+		{input: `hello \"world\"`, want: `hello "world"`},
+		{input: `\n`, want: `\n`},
+		{input: `no escapes here`, want: `no escapes here`},
+	}
+	for _, tt := range tests {
+		t.Run(tt.input, func(t *testing.T) {
+			if got := UnescapeQuotedString(tt.input); got != tt.want {
+				t.Errorf("UnescapeQuotedString(%q) = %q, want %q", tt.input, got, tt.want)
+			}
+		})
+	}
+}
+
 func TestRandomStringConcurrency(t *testing.T) {
 	// Make sure random strings don't crash under high concurrency.
-	for i := 0; i < 5000; i++ {
+	for range 5000 {
 		go RandomString(10000)
 	}
 }
