@@ -21,6 +21,7 @@ package auditlog
 
 import (
 	"fmt"
+	"net/http"
 	"strings"
 
 	"github.com/corazawaf/coraza/v3/experimental/plugins/plugintypes"
@@ -45,81 +46,102 @@ func (nativeFormatter) Format(al plugintypes.AuditLog) ([]byte, error) {
 		res.WriteString(boundaryPrefix)
 		res.WriteByte(byte(part))
 		res.WriteString("--\n")
-		// [27/Jul/2016:05:46:16 +0200] V5guiH8AAQEAADTeJ2wAAAAK 192.168.3.1 50084 192.168.3.111 80
-		_, _ = fmt.Fprintf(&res, "[%s] %s %s %d %s %d", al.Transaction().Timestamp(), al.Transaction().ID(),
-			al.Transaction().ClientIP(), al.Transaction().ClientPort(), al.Transaction().HostIP(), al.Transaction().HostPort())
+
+		addSeparator := true
+
 		switch part {
+		case types.AuditLogPartHeader:
+			// Part A: Audit log header containing only the timestamp and transaction info line
+			// Note: Part A does not have an empty line separator after it
+			_, _ = fmt.Fprintf(&res, "[%s] %s %s %d %s %d\n",
+				al.Transaction().Timestamp(), al.Transaction().ID(),
+				al.Transaction().ClientIP(), al.Transaction().ClientPort(),
+				al.Transaction().HostIP(), al.Transaction().HostPort())
+			addSeparator = false
 		case types.AuditLogPartRequestHeaders:
-			// GET /url HTTP/1.1
-			// Host: example.com
-			// User-Agent: Mozilla/5.0
-			// Accept: text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8
-			// Accept-Language: en-US,en;q=0.5
-			// Accept-Encoding: gzip, deflate
-			// Referer: http://example.com/index.html
-			// Connection: keep-alive
-			// Content-Type: application/x-www-form-urlencoded
-			// Content-Length: 6
-			_, _ = fmt.Fprintf(
-				&res,
-				"\n%s %s %s",
-				al.Transaction().Request().Method(),
-				al.Transaction().Request().URI(),
-				al.Transaction().Request().Protocol(),
-			)
-			for k, vv := range al.Transaction().Request().Headers() {
-				for _, v := range vv {
-					res.WriteByte('\n')
-					res.WriteString(k)
-					res.WriteString(": ")
-					res.WriteString(v)
+			// Part B: Request headers
+			if al.Transaction().HasRequest() {
+				_, _ = fmt.Fprintf(
+					&res,
+					"%s %s %s",
+					al.Transaction().Request().Method(),
+					al.Transaction().Request().URI(),
+					al.Transaction().Request().Protocol(),
+				)
+				for k, vv := range al.Transaction().Request().Headers() {
+					for _, v := range vv {
+						res.WriteByte('\n')
+						res.WriteString(k)
+						res.WriteString(": ")
+						res.WriteString(v)
+					}
 				}
+				res.WriteByte('\n')
 			}
 		case types.AuditLogPartRequestBody:
-			if body := al.Transaction().Request().Body(); body != "" {
-				res.WriteByte('\n')
-				res.WriteString(body)
+			// Part C: Request body
+			if al.Transaction().HasRequest() {
+				if body := al.Transaction().Request().Body(); body != "" {
+					res.WriteString(body)
+					res.WriteByte('\n')
+				}
 			}
 		case types.AuditLogPartIntermediaryResponseBody:
-			if body := al.Transaction().Response().Body(); body != "" {
-				res.WriteByte('\n')
-				res.WriteString(al.Transaction().Response().Body())
+			// Part E: Intermediary response body
+			if al.Transaction().HasResponse() {
+				if body := al.Transaction().Response().Body(); body != "" {
+					res.WriteString(body)
+					res.WriteByte('\n')
+				}
 			}
 		case types.AuditLogPartResponseHeaders:
-			for k, vv := range al.Transaction().Response().Headers() {
-				for _, v := range vv {
-					res.WriteByte('\n')
-					res.WriteString(k)
-					res.WriteString(": ")
-					res.WriteString(v)
+			// Part F: Response headers
+			if al.Transaction().HasResponse() {
+				// Write status line: HTTP/1.1 200 OK
+				protocol := al.Transaction().Response().Protocol()
+				if protocol == "" {
+					protocol = "HTTP/1.1"
+				}
+				status := al.Transaction().Response().Status()
+				statusText := http.StatusText(status)
+				_, _ = fmt.Fprintf(&res, "%s %d %s\n", protocol, status, statusText)
+
+				// Write headers
+				for k, vv := range al.Transaction().Response().Headers() {
+					for _, v := range vv {
+						res.WriteString(k)
+						res.WriteString(": ")
+						res.WriteString(v)
+						res.WriteByte('\n')
+					}
 				}
 			}
 		case types.AuditLogPartAuditLogTrailer:
-			// Stopwatch: 1470025005945403 1715 (- - -)
-			// Stopwatch2: 1470025005945403 1715; combined=26, p1=0, p2=0, p3=0, p4=0, p5=26, ↩
-			// sr=0, sw=0, l=0, gc=0
-			// Response-Body-Transformed: Dechunked
-			// Producer: ModSecurity for Apache/2.9.1 (http://www.modsecurity.org/).
-			// Server: Apache
-			// Engine-Mode: "ENABLED"
-
-			// AuditLogTrailer is also expected to contain the error message generated by the rule, if any
+			// Part H: Audit log trailer
 			for _, alEntry := range al.Messages() {
 				alWithErrMsg, ok := alEntry.(auditLogWithErrMesg)
 				if ok && alWithErrMsg.ErrorMessage() != "" {
-					res.WriteByte('\n')
 					res.WriteString(alWithErrMsg.ErrorMessage())
+					res.WriteByte('\n')
 				}
 			}
-
-			_, _ = fmt.Fprintf(&res, "\nStopwatch: %s\nResponse-Body-Transformed: %s\nProducer: %s\nServer: %s", "", "", "", "")
 		case types.AuditLogPartRulesMatched:
+			// Part K: Matched rules
 			for _, alEntry := range al.Messages() {
-				res.WriteByte('\n')
 				res.WriteString(alEntry.Data().Raw())
+				res.WriteByte('\n')
 			}
+		case types.AuditLogPartEndMarker:
+			// Part Z: Final boundary marker with no content
+		default:
+			// For any other parts (D, G, I, J) that aren't explicitly handled,
+			// they remain empty
 		}
-		res.WriteByte('\n')
+
+		// Add separator newline for all parts except A
+		if addSeparator {
+			res.WriteByte('\n')
+		}
 	}
 
 	return []byte(res.String()), nil
