@@ -991,7 +991,7 @@ func (tx *Transaction) ReadRequestBodyFrom(r io.Reader) (*types.Interruption, in
 // ProcessRequestBody Performs the analysis of the request body (if any)
 //
 // It is recommended to call this method even if it is not expected to have a body.
-// It permits to execute rules belonging to request body phase, but not necesarily
+// It permits to execute rules belonging to request body phase, but not necessarily
 // processing the request body.
 //
 // Remember to check for a possible intervention.
@@ -1241,7 +1241,7 @@ func (tx *Transaction) ReadResponseBodyFrom(r io.Reader) (*types.Interruption, i
 // ProcessResponseBody Perform the analysis of the the response body (if any)
 //
 // It is recommended to call this method even if it is not expected to have a body.
-// It permits to execute rules belonging to request body phase, but not necesarily
+// It permits to execute rules belonging to request body phase, but not necessarily
 // processing the response body.
 //
 // note Remember to check for a possible intervention.
