@@ -1,8 +1,7 @@
+//go:build !tinygo
+
 // Copyright 2024 Juan Pablo Tosso and the OWASP Coraza contributors
 // SPDX-License-Identifier: Apache-2.0
-//go:build !tinygo
-// +build !tinygo
-
 package operators
 
 import (
