@@ -3,7 +3,6 @@
 
 // These benchmarks don't currently compile with TinyGo
 //go:build !tinygo
-// +build !tinygo
 
 package coreruleset
 
@@ -205,6 +204,8 @@ SecRule REQUEST_HEADERS:X-CRS-Test "@rx ^.*$" \
 	if err != nil {
 		t.Fatalf("failed to create error log: %v", err)
 	}
+	defer errorFile.Close()
+
 	errorWriter := bufio.NewWriter(errorFile)
 	conf = conf.WithErrorCallback(func(rule types.MatchedRule) {
 		msg := rule.ErrorLog() + "\n"
